add separate warning flags for the library: -Wpedantic -Wold-style-cast -Wsign-conversion

+++ cpputest-for-qpcpp-lib/src/cms_cpputest_qf_ctrl.cpp
@@ -69,7 +69,7 @@ void Setup(enum_t const maxPubSubSignalValue, uint32_t ticksPerSecond,
     l_memPoolOption     = memPoolOpt;
     l_ticksPerSecond    = ticksPerSecond;
     l_subscriberStorage = new SubscriberList();
-    l_subscriberStorage->resize(maxPubSubSignalValue);
+    l_subscriberStorage->resize(static_cast<size_t>(maxPubSubSignalValue));
     QSubscrList nullValue = QSubscrList();
     std::fill(l_subscriberStorage->begin(), l_subscriberStorage->end(),
               nullValue);
